+++ bin/pathway_analysis.R
@@ -4,17 +4,19 @@
 # Author: Gisela Gabernet
 # QBiC 2019; MIT License
 
-library(gprofiler2)
-library(ggplot2)
-library(reshape2)
-library(pheatmap)
-library(pathview)
-library(AnnotationDbi)
-library(optparse)
+invisible( lapply(c(
+"gprofiler2",
+"ggplot2",
+"reshape2",
+"pheatmap",
+"pathview",
+"AnnotationDbi",
+"optparse",
+"org.Mm.eg.db",
+"org.Hs.eg.db"
+), library, character.only=T))
+# Need to load library for your species! Or provide as parameter to biocmanager
 
-# Need to load library for your species
-library(org.Mm.eg.db) #Mmusculus TODO: Delete these from container? If fast!
-library(org.Hs.eg.db) #Hsapiens
 
 # ------------------
 # Reading parameters
@@ -177,13 +179,6 @@ for (file in contrast_files){
                     file = paste0(outdir, "/", fname, "/", fname, "_pathway_enrichment_results.tsv"),
                     sep="\t", quote = F, col.names = T, row.names = F)
 
-        # TODO debugging summary table
-        # Collecting summary variables
-        # contrast <- append(contrast, fname)
-        # number_DE_genes <- append(number_DE_genes, length(DE_genes$Ensembl_ID))
-        # number_enriched_pathways <- append(number_enriched_pathways, summary(as.factor(pathway_gostres_table$source)))
-        # list_DE_genes <- append(list_DE_genes, DE_genes$Ensembl_ID)
-        # list_enriched_pathways <- append(list_enriched_pathways, pathway_gostres_table$term_id)
 
         # Printing summary variables
         print("------------------------------------")